Reuse accuracy form better
