+++ app/views/payments/accuracy.html.slim
@@ -1,25 +1,3 @@
 h2 Payment details
 
-= form_for @form, as: :payment, url: :accuracy_save_payment, method: :post, html: { autocomplete: 'off', id: 'accuracy-form' } do |f|
-  .row
-    .small-12.medium-8.large-5.columns
-      .form-group
-        .row.collapse
-          .columns.small-12
-            = f.label :correct, t('payments.labels.correct')
-            .options.radio
-              .option
-                label for='payment_correct_false'
-                  = f.radio_button :correct, 'false'
-                  = t('payments.incorrect')
-              .option
-                label for='payment_correct_true'
-                  = f.radio_button :correct, 'true'
-                  = t('payments.correct')
-
-      .panel-indent.form-group.row.collapse#reason-input
-        .columns.small-12
-          = f.label :incorrect_reason, t('payments.labels.incorrect_reason')
-          = f.text_area :incorrect_reason, rows: 3
-
-  = f.submit 'Next', :class => 'button primary'
+= render 'shared/forms/accuracy', { form: @form, model: :payment, url: :accuracy_save_payment }
